+++ frontend/src/common/components/login/index.tsx
@@ -1,7 +1,7 @@
 import { FC, useState } from "react";
 import { SubmitHandler, useForm } from "react-hook-form";
 import { useNavigate } from "react-router";
-import { useRegisterMutation } from "../store/api/registerApi";
+import { useLoginMutation } from "../../../store/api/authApi";
 
 type IFormInput = {
   firstName: string;
@@ -9,36 +9,31 @@ type IFormInput = {
   password: string;
 };
 
-export const RegisterPage: FC = () => {
+export const LoginPage: FC = () => {
   const navigate = useNavigate();
   const goToHomePage = () => navigate("/");
-  const goToLoginPage = () => navigate("/login");
+  const goToRegisterPage = () => navigate("/register");
   const goToAuthPage = () => navigate("/auth");
-
   const [error, setError] = useState<string | null>(null);
 
-  const [mutation] = useRegisterMutation();
+  const [mutation] = useLoginMutation();
 
   const { register, handleSubmit } = useForm<IFormInput>();
   const onSubmit: SubmitHandler<IFormInput> = async (data) => {
     try {
-      await mutation({
-        firstName: data.firstName,
-        email: data.email,
-        password: data.password,
-      }).unwrap();
+      await mutation({ email: data.email, password: data.password }).unwrap();
+      console.log("ログイン成功");
       goToAuthPage();
     } catch (error: any) {
       setError(error?.data?.error || "予期せぬエラーが発生しました");
     }
   };
+
   return (
     <>
-      {error && <p>{error}</p>}
-      <h1>登録ページ</h1>
+      <p>{error}</p>
+      <h1>ログインページ</h1>
       <form onSubmit={handleSubmit(onSubmit)}>
-        <label>名前</label>
-        <input {...register("firstName")} />
         <label>メールアドレス</label>
         <input {...register("email")} />
         <label>パスワード</label>
@@ -46,7 +41,7 @@ export const RegisterPage: FC = () => {
         <input type="submit" />
       </form>
       <button onClick={goToHomePage}>ホームページ</button>
-      <button onClick={goToLoginPage}>ログインページ</button>
+      <button onClick={goToRegisterPage}>登録ページ</button>
     </>
   );
 };
